Change test

+++ src/main/java/ru/netolgy/stats/StatsService.java
@@ -1,21 +1,21 @@
 package ru.netolgy.stats;
 
 public class StatsService {
-    public int sumSales(int[] sales) {
-        int sumSales = 0;  // начальное значение суммы продаж
+    public long sumSales(long[] sales) {
+        long sumSales = 0;  // начальное значение суммы продаж
         for (int i = 0; i < sales.length; i++) {
             sumSales = sumSales + sales[i];
         }
         return sumSales;
     }
 
-    public float averageSales(int[] sales) {
-        int sumSales = sumSales(sales);  // сумма продаж за год
-        float averageSales = sumSales / sales.length; // средняя сумма продаж за месяц
+    public long averageSales(long[] sales) {
+        long sumSales = sumSales(sales);  // сумма продаж за год
+        long averageSales = sumSales / sales.length; // средняя сумма продаж за месяц
         return averageSales;
     }
 
-    public int maxSales(int[] sales) {
+    public int maxSales(long[] sales) {
         int maxMonth = 0; // номер месяца с максимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -27,7 +27,7 @@ public class StatsService {
         return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int minSales(int[] sales) {
+    public int minSales(long[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -39,9 +39,9 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public int aboveAverageSales(int[] sales) {
+    public int aboveAverageSales(long[] sales) {
         int aboveMonth = 0; // количество месяцев с продажами выше средних за месяц
-        float averageSales = averageSales(sales);
+        long averageSales = averageSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] > averageSales) { // значит, в рассматриваемом i-м месяце продажи выше средних
@@ -52,9 +52,9 @@ public class StatsService {
         return aboveMonth;
     }
 
-    public int belowAverageSales(int[] sales) {
+    public int belowAverageSales(long[] sales) {
         int belowMonth = 0; // количество месяцев с продажами ниже средних за месяц
-        float averageSales = averageSales(sales);
+        long averageSales = averageSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < averageSales) { // значит, в рассматриваемом i-м месяце продажи ниже средних
